python-service/requirements.txt
flask
flask-cors
openai-whisper
librosa
mediapipe
opencv-python
soundfile
